chore: Update .gitignore and HttpInstance.dart
This commit updates the .gitignore file to include the .env file, ensuring it is not tracked by Git. It also modifies the HttpInstance.dart file to use the BACKEND_URL environment variable from the .env file, allowing for dynamic configuration of the base URL. These changes improve the organization and flexibility of the codebase.